requirements.txt
aiogram==2.25.1
psycopg2-binary
python-dotenv==1.0.0
SQLAlchemy==2.0.9
nest_asyncio
requests